+++ lib/tools/backgroundsync.dart
@@ -68,7 +68,7 @@ class BackgroundSync {
     }
   }
 
-  static Future<void> executeSync() async {
+  static Future<void> executeSync({bool fromScan = false}) async {
     //this static method can't access the providers we already have so we have to re-invent some things here...
     Uint8List _encryptionKey;
     var _secureStorage = const FlutterSecureStorage();
@@ -136,7 +136,9 @@ class BackgroundSync {
               //addr does not have a pending notification
               adressesToQuery[walletAddress.address] = wallet.transactions
                   .where(
-                    (element) => element.address == walletAddress.address,
+                    (element) =>
+                        element.address == walletAddress.address &&
+                        element.direction == 'in',
                   )
                   .length;
             }
@@ -160,6 +162,7 @@ class BackgroundSync {
           if (result.body.contains('foundDifference')) {
             //valid answer
             var bodyDecoded = jsonDecode(result.body);
+            print(bodyDecoded);
             _foundDifference = bodyDecoded['foundDifference'];
             if (_foundDifference == true) {
               //loop through addresses in result
@@ -169,7 +172,13 @@ class BackgroundSync {
                 wallet.putPendingTransactionNotification(PendingNotification(
                     address: element['address'], tx: element['tx']));
               });
-              _shouldNotify = true;
+
+              if (fromScan == true) {
+                //persist backend data
+                wallet.clearPendingTransactionNotifications();
+              } else {
+                _shouldNotify = true;
+              }
             }
           }